requirements.txt
# AI Real Estate Valuation System - Dependencies
# Install with: pip install -r requirements.txt

# Web framework
streamlit>=1.32.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0,<2.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.13.0

# Database
psycopg2-binary>=2.9.9

# AI / LLM
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0
